% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asypeer.estim.R
\name{asypeer.estim}
\alias{asypeer.estim}
\title{Estimating the Asymmetric Peer Effects Model}
\usage{
asypeer.estim(
  formula,
  excluded.instruments,
  Glist,
  data,
  common.gamma,
  spillover = FALSE,
  asymmetry = TRUE,
  weight = "IV",
  HAC = "group-iid",
  nboot = 500,
  fixed.effects = FALSE,
  tol = 1e-10,
  nthread = 1,
  drop = NULL,
  tol.optim = .Machine$double.eps^0.25,
  ...
)
}
\arguments{
\item{formula}{An object of class \link[stats]{formula}: a symbolic description of the model.
\code{formula} should be specified as \code{y ~ x1 + x2 + ....}, where \code{y} is the outcome and \code{x1}, \code{x2}, ....
are the vectors or matrices of control variables, which may include contextual variables such as averages among peers.}

\item{excluded.instruments}{A \link[stats]{formula} specifying the excluded
instruments. It should be written as \code{~ z1 + z2 + ...}, where \code{z1},
\code{z2}, ... are the vectors or matrices of excluded instruments for the two
endogenous variables in the asymmetric model: the average peers' outcomes and
the average difference in outcomes between an agent and her peers who have a
higher outcome. If omitted, default instruments are generated using
\link{gen.instrument}.}

\item{Glist}{The adjacency matrix. For networks consisting of multiple subnets (e.g., schools),
\code{Glist} must be a list of subnets, with the \code{s}-th element being an \eqn{n_s \times n_s}
adjacency matrix, where \eqn{n_s} is the number of nodes in the \code{s}-th subnet.}

\item{data}{An optional data frame, list, or environment (or an object that can be coerced to a
data frame via \link[base]{as.data.frame}) containing the variables in the model. If a variable
is not found in \code{data}, it is taken from \code{environment(formula)}, typically the environment
from which \code{asypeer.estim} is called.}

\item{common.gamma}{A character vector indicating the control variables assumed to have the same marginal effect on marginal utility for isolated and non-isolated agents.
If omitted, all control variables are included. A value of \code{NULL} or \code{character(0)} indicates that none of the variables.}

\item{spillover}{A logical value indicating if the model included a spillover component in additional to the asymmetric conformity effect.}

\item{asymmetry}{A logical value indicating if the preference for conformity is asymmetric or not.}

\item{weight}{A character string specifying the weighting matrix used in the GMM estimation.
Available options are: \code{"identity"} for GMM with the identity matrix as the weighting matrix;
\code{"IV"} for the standard instrumental-variable GMM estimator; and \code{"optimal"} for GMM with the
optimal weighting matrix.}

\item{HAC}{A character string specifying the correlation structure of the idiosyncratic errors
used for covariance estimation. Options are \code{"iid"} for independent errors; \code{"group iid"} for
independence within the groups of isolated and non-isolated players; \code{"hetero"} for
heteroskedastic but non-autocorrelated errors; \code{"cluster"} for heteroskedastic errors with
potential within-subnetwork correlation; and \code{"cboot"} for the pair-bootstrap version of
\code{"cluster"}.}

\item{nboot}{The number of bootstrap simulations when \code{HAC = "cluster"}.}

\item{fixed.effects}{A logical value indicating whether the model includes subnetwork fixed effects.}

\item{tol}{A tolerance value used in the QR factorization to detect collinear columns in the
matrices of explanatory variables and instruments, ensuring a full-rank matrix (see
\link[base]{qr}).}

\item{nthread}{Number of CPU cores (threads) used to run parts of the estimation in parallel.}

\item{drop}{A dummy vector of the same length as the sample, indicating whether an observation should be dropped.
This can be used, for example, to remove false isolates or to estimate the model only on non-isolated agents.
These observations cannot be directly removed from the network by the user because they may still be friends with other agents.}

\item{tol.optim}{The tolerance value for \link[stats]{optimize}.
The objective function is concentrated in a
function of \eqn{\beta_l} or \eqn{\beta} alone, which is optimized
using \link[stats]{optimize}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A list containing:
\item{model.info}{A list with information about the model, such as the number of subnets, number of observations, and other key details.}
\item{gmm}{A list of GMM estimation results, including parameter estimates, the covariance matrix, and related statistics.}
\item{data}{A list including the original data used to estimate the model, as well as the endogenous variables and their instruments.}
}
\description{
\code{asypeer.estim} estimates the asymmetric peer effects model. The instruments are generated automatically following the procedure described
in the paper and implemented in \link{gen.instruments}. The user can also supply her own instruments.
}
\examples{
if (requireNamespace("PartialNetwork", quietly = TRUE)) {
library(PartialNetwork)
ngr  <- 50  # Number of subnets
nvec <- rep(30, ngr)  # Size of subnets
n    <- sum(nvec)

### Simulating Data
## Network matrix
G   <- lapply(1:ngr, function(z) {
 Gz <- matrix(rbinom(nvec[z]^2, 1, 0.3), nvec[z], nvec[z])
 diag(Gz) <- 0
 # Adding isolated nodes (important for the structural model)
 niso <- sample(0:nvec[z], 1, prob = ((nvec[z] + 1):1)^5 / sum(((nvec[z] + 1):1)^5))
 if (niso > 0) {
   Gz[sample(1:nvec[z], niso), ] <- 0
 }
 Gz
})

Gnorm   <- norm.network(G)
X       <- cbind(rnorm(n, 0, 2), rpois(n, 2))
GX      <- peer.avg(Gnorm, X)
delta   <- 0.25
beta    <- c(0.3, 0.6)
gamma   <- c(4, 1, -0.7, 0, -0.5) 
eps     <- rnorm(n, 0, 0.5) 

## Generating `y`
y <- asypeer.sim(formula = ~ X + GX, Glist = Gnorm, delta = delta, beta = beta, 
                gamma = gamma, epsilon = eps)
y <- y$y

### Estimating the asymmetric peer effects model
est <- asypeer.estim(formula=y ~ X + GX, Glist = Gnorm, spillover = TRUE)
summary(est, diagnostic = TRUE)
}                  
}
