% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asypeer.sim.R
\name{asypeer.sim}
\alias{asypeer.sim}
\title{Simulating Asymmetric Peer Effect Models with Continuous Outcome}
\usage{
asypeer.sim(
  formula,
  Glist,
  parms,
  beta,
  delta,
  gamma,
  epsilon,
  init,
  tol = 1e-10,
  maxit = 500,
  nthread = 1,
  print = FALSE,
  data
)
}
\arguments{
\item{formula}{A formula object (\link[stats]{formula}): a symbolic description of the model. \code{formula} should be specified as, for example, \code{~ x1 + x2},
where \code{x1} and \code{x2} are control variables, which can include contextual variables such as averages or quantiles among peers.}

\item{Glist}{The adjacency matrix. For networks consisting of multiple subnets (e.g., schools), \code{Glist} must be a list of subnets, with the \code{s}-th element being an \eqn{n_s \times n_s} adjacency matrix, where \eqn{n_s} is the number of nodes in the \code{s}-th subnet.}

\item{parms}{A vector defining the true values of \eqn{(\boldsymbol{\beta}', \delta, \boldsymbol{\gamma}')'}, where \eqn{\boldsymbol{\beta} = (\beta^l, \beta^h)'} captures asymmetric conformity effects,
\eqn{\delta} measures spillover effects, and \eqn{\boldsymbol{\gamma}} is the parameter vector associated with the exogenous characteristics in agent types (i.e., the control variables \code{x1}, \code{x2}, ect.).
The parameters \eqn{\delta}, \eqn{\boldsymbol{\beta}}, and \eqn{\boldsymbol{\gamma}} can also be specified separately using the arguments \code{delta}, \code{beta}, and \code{gamma}.}

\item{beta}{The true value of the asymmetric conformity parameters.}

\item{delta}{The true value of the spillover effect parameter.}

\item{gamma}{The true value of the vector \eqn{\boldsymbol{\gamma}}.}

\item{epsilon}{A vector of idiosyncratic error terms. If not specified, it will be simulated from a standard normal distribution.}

\item{init}{An optional initial guess for the equilibrium.}

\item{tol}{The tolerance value used in the Fixed Point Iteration Method to compute the outcome \code{y}. The process stops if the \eqn{\ell_1}-distance
between two consecutive values of \code{y} is less than \code{tol}.}

\item{maxit}{The maximum number of iterations for the Fixed Point Iteration Method.}

\item{nthread}{Number of CPU cores (threads) used to run parts of the simulation in parallel.}

\item{print}{A logical value indicating whether the \eqn{\ell_1} distance at each iteration of the best‑response dynamics should be printed.}

\item{data}{An optional data frame, list, or environment containing the model variables. If a variable is not found in \code{data}, it is retrieved from \code{environment(formula)}, typically the environment from which \code{asypeer.sim} is called.}
}
\value{
A list containing:
\item{y}{The simulated variable.}
\item{epsilon}{The idiosyncratic error.}
\item{init}{The initial guess.}
\item{iteration}{The number of iterations before convergence.}
}
\description{
\code{asypeer.sim} simulates asymmetric peer effect models with continuous responses.
}
\examples{
\donttest{
if (requireNamespace("PartialNetwork", quietly = TRUE)) {
library(PartialNetwork)
ngr  <- 50  # Number of subnets
nvec <- rep(30, ngr)  # Size of subnets
n    <- sum(nvec)

### Simulating Data
## Network matrix
G <- lapply(1:ngr, function(z) {
 Gz       <- matrix(rbinom(nvec[z]^2, 1, 0.3), nvec[z], nvec[z])
 diag(Gz) <- 0
 # Adding isolated nodes (important for the structural model)
 niso <- sample(0:nvec[z], 1, prob = ((nvec[z] + 1):1)^5 / sum(((nvec[z] + 1):1)^5))
 if (niso > 0) {
   Gz[sample(1:nvec[z], niso), ] <- 0
 }
 Gz
})

Gnorm   <- norm.network(G)
X       <- cbind(rnorm(n, 0, 2), rpois(n, 2))
GX      <- peer.avg(Gnorm, X)
delta   <- 0.25
beta    <- c(0.3, 0.6)
gamma   <- c(4, 1, -0.7, 0, -0.5) 
eps     <- rnorm(n, 0, 0.5) 

#' ## Generating `y`
y <- asypeer.sim(formula = ~ X + GX, Glist = Gnorm, delta = delta, beta = beta, 
                gamma = gamma, epsilon = eps)}
}                
}
