% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peer.asyavg.R
\name{peer.asyavg}
\alias{peer.asyavg}
\title{Computing peer (asymmetric) average values}
\usage{
peer.asyavg(formula, Glist, data, nthread = 1)
}
\arguments{
\item{formula}{An object of class \link[stats]{formula}, which should be specified as \code{y ~ X},
where \code{y} is the dependent variable used to characterized which friends have a higher or lower outcome
and \code{X} is the vector or matrix of variables whose average values among peers should be computed.}

\item{Glist}{The adjacency matrix. For networks consisting of multiple subnets (e.g., schools),
\code{Glist} must be a list of subnets, with the \code{m}-th element being an \eqn{n_m \times n_m}
adjacency matrix, where \eqn{n_m} is the number of nodes in the \code{m}-th subnet.}

\item{data}{An optional data frame, list, or environment (or an object that can be coerced to a
data frame via \link[base]{as.data.frame}) containing the variables in the model. If a variable
is not found in \code{data}, it is taken from \code{environment(formula)}, typically the environment
from which \code{peer.asyavg} is called.}

\item{nthread}{A strictly positive integer specifying the number of threads used in
computationally intensive steps.}
}
\value{
A list containing:
\item{degree}{The row sum of the network matrix. As \code{Glist} is row-normalized, it is either \code{0} or \code{1}.}
\item{peer.avg}{The average of \code{X} (see the \code{formula} argument) among all peers.}
\item{hdegree}{The share of higher-performing friends.}
\item{hpeer.avg}{The sum of \code{X} weighted by \eqn{g_{ij}} among higher-performing friends.}
\item{ldegree}{The share of lower-performing friends.}
\item{lpeer.avg}{The sum of \code{X} weighted by \eqn{g_{ij}} among lower-performing friends.}
}
\description{
\code{peer.asyavg} computes the average values of a vector or matrix \code{X} among friends,
differentiating between friends with a higher or lower outcome than the agent.
}
\examples{
if (requireNamespace("PartialNetwork", quietly = TRUE)) {
library(PartialNetwork)
ngr  <- 50  # Number of subnets
nvec <- rep(30, ngr)  # Size of subnets
n    <- sum(nvec)

### Simulating Data
## Network matrix
G <- lapply(1:ngr, function(z) {
  Gz       <- matrix(rbinom(nvec[z]^2, 1, 0.3), nvec[z], nvec[z])
  diag(Gz) <- 0
  # Adding isolated nodes (important for the structural model)
  niso <- sample(0:nvec[z], 1, prob = ((nvec[z] + 1):1)^5 / sum(((nvec[z] + 1):1)^5))
  if (niso > 0) {
    Gz[sample(1:nvec[z], niso), ] <- 0
  }
  Gz
})
Gnorm   <- norm.network(G)
X       <- cbind(rnorm(n, 0, 2), rpois(n, 2))
GX      <- peer.avg(Gnorm, X)
colnames(X)<-c("X1","X2") 
delta   <- 0.25
beta    <- c(0.3, 0.6)
gamma   <- c(4, 1, -0.7, 0, -0.5) 
eps     <- rnorm(n, 0, 0.5) 

## Generating `y`
y <- asypeer.sim(formula = ~ X + GX, Glist = Gnorm, delta = delta, beta = beta, 
                gamma = gamma, epsilon = eps, nthread = 5)
y <- y$y

### Computing averages X among peers
peeravg <- peer.asyavg(formula = y ~ X, Glist = Gnorm, nthread = 1)
}
}
