% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spillover.R
\name{spillover}
\alias{spillover}
\title{Find the optimal targeting order in an intervention}
\usage{
spillover(
  asymodel,
  symodel,
  Glist,
  targ.net,
  data,
  treatment = 1,
  nthread = 1,
  print = TRUE,
  tol = 1e-12
)
}
\arguments{
\item{asymodel}{An object of class \code{\link{asypeer.estim}} or
\code{\link{summary.asypeer.estim}},
estimated using the asymmetric specification (\code{asymmetry=TRUE}).}

\item{symodel}{An object of class \code{\link{asypeer.estim}} or
\code{\link{summary.asypeer.estim}},
estimated using the symmetric specification (\code{asymmetry=FALSE}).}

\item{Glist}{The adjacency matrix or list of adjacency matrices. For networks
consisting of multiple subnets (e.g., schools), \code{Glist} must be a list,
where the \code{s}-th element is an \eqn{n_s x n_s} adjacency matrix and
\eqn{n_s} is the number of nodes in subnet \code{s}.}

\item{targ.net}{A scalar indicating the index in \code{Glist} of the network
for which the targeting order and spillovers must be computed.}

\item{data}{An optional data frame, list, or environment (or an object that can
be coerced to a data frame via \link[base]{as.data.frame}) containing the
variables in the model. If a variable is not found in \code{data}, it is
retrieved from the environment of \code{asymodel}.}

\item{treatment}{A scalar, or a vector whose length equals the
number of agents in the targeted network, indicating the value of the treatment.}

\item{nthread}{The number of CPU cores (threads) used for parallel computation.}

\item{print}{A logical value indicating whether a progress bar should be printed.}

\item{tol}{A numeric tolerance used to assess the convergence, post-intervention,
to a Nash equilibrium}
}
\value{
A list containing:
\item{targeted}{A matrix or list containing the indices
of the intervention's targets, ordered by their influence.}
\item{budget}{A matrix indicating the intervention's budget for any number
of targeted agents.}
\item{diff.y}{A matrix indicating the difference between the baseline welfare
and the welfare post-intervention, for any number of targeted agents.}
\item{spillover}{A matrix giving the spillover associated
with each targeted agent.}
}
\description{
\code{spillover} identifies the optimal targeting order in an intervention aiming at
maximizing the utilitarian welfare and computes the associated individual spillovers.
The targeting order and the spillovers can be generated via different targeting methods:
using the symmetric specification or the asymmetric specification. In the latter case,
the optimal order is computed by forward or backward optimization.
}
\examples{
\donttest{
if (requireNamespace("PartialNetwork", quietly = TRUE)) {
  library(PartialNetwork)
  ngr  <- 50  # Number of subnets
  nvec <- rep(30, ngr)  # Size of subnets
  n    <- sum(nvec)
  
  ### Simulating Data
  ## Network matrix
  G   <- lapply(1:ngr, function(z) {
    Gz <- matrix(rbinom(nvec[z]^2, 1, 0.3), nvec[z], nvec[z])
    diag(Gz) <- 0
    # Adding isolated nodes (important for the structural model)
    niso <- sample(0:nvec[z], 1, prob = ((nvec[z] + 1):1)^5 / sum(((nvec[z] + 1):1)^5))
    if (niso > 0) {
      Gz[sample(1:nvec[z], niso), ] <- 0
    }
    Gz
  })
  
  Gnorm   <- norm.network(G)
  X       <- cbind(rnorm(n, 0, 2), rpois(n, 2))
  GX      <- peer.avg(Gnorm, X)
  delta   <- 0.25
  beta    <- c(0.3, 1.6)
  gamma   <- c(4, 1, -0.7, 0, -0.5)
  eps     <- rnorm(n, 0, 0.5)
  
  ## Generating `y`
  y <- asypeer.sim(formula = ~ X + GX, Glist = Gnorm, delta = delta, beta = beta,
                   gamma = gamma, epsilon = eps)
  y <- y$y
  
  ### Estimating a symmetric peer effects model
  mod1 <- asypeer.estim(formula = y ~ X + GX, Glist = Gnorm, spillover = TRUE,
                        asymmetry = FALSE)
  summary(mod1)
  
  ### Estimating an asymmetric peer effects model
  mod2 <- asypeer.estim(formula = y ~ X + GX, Glist = Gnorm, spillover = TRUE)
  summary(mod2)
  
  ## treatment
  treat <- spillover(asymodel = mod2, symodel = mod1, Glist = Gnorm, targ.net = 1,
                     treatment = 1)
  
  ## Plot
  plot(treat)
}}
}
