% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peer.ces.R
\name{summary.cesconfpeer}
\alias{summary.cesconfpeer}
\alias{print.summary.cesconfpeer}
\alias{print.cesconfpeer}
\title{Summary for the Estimation of CES-based Peer Effects Models}
\usage{
\method{summary}{cesconfpeer}(object, fullparameters = TRUE, ...)

\method{print}{summary.cesconfpeer}(x, ...)

\method{print}{cesconfpeer}(x, ...)
}
\arguments{
\item{object}{An object of class \code{\link{cesconfpeer}}.}

\item{fullparameters}{A logical value indicating whether all parameters should be summarized (may be useful for the structural model).}

\item{...}{Further arguments passed to or from other methods.}

\item{x}{An object of class \code{\link{summary.cesconfpeer}} or \code{\link{cesconfpeer}}.}
}
\value{
A list containing:
\item{model.info}{A list with information about the model, such as the number of subnets, number of observations, and other key details.}
\item{coefficients}{A summary of the estimates, standard errors, and p-values.}
\item{gmm}{A list of GMM estimation results, including parameter estimates, the covariance matrix, and related statistics.}
}
\description{
Summary and print methods for the class \code{\link{cesconfpeer}}.
}
