\name{AsyPeer-package}
\alias{AsyPeer-package}
\alias{AsyPeer}
\docType{package}
\title{
\packageTitle{AsyPeer}
}
\description{
\packageDescription{AsyPeer}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{AsyPeer}

Maintainer: \packageMaintainer{AsyPeer}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = AsyPeer)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{AsyPeer}}
%\section{Documentation Index}{\packageIndices{AsyPeer}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
