% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asypeer.estim.R
\name{asypeer.estim}
\alias{asypeer.estim}
\title{Estimate the Asymmetric Peer Effects Model}
\usage{
asypeer.estim(
  formula,
  excluded.instruments,
  Glist,
  data,
  common.gamma,
  spillover = FALSE,
  asymmetry = TRUE,
  weight = "IV",
  HAC = "group-iid",
  nboot = 500,
  fixed.effects = FALSE,
  tol = 1e-10,
  nthread = 1,
  drop = NULL,
  tol.optim = .Machine$double.eps^0.25,
  gen.inst.arg = list(),
  ...
)
}
\arguments{
\item{formula}{An object of class \link[stats]{formula}: a symbolic description
of the model. It should be specified as \code{y ~ x1 + x2 + ...}, where
\code{y} is the outcome variable and \code{x1}, \code{x2}, ... are control
variables, which may include contextual variables such as peer averages.}

\item{excluded.instruments}{A \link[stats]{formula} specifying the excluded
instruments. It should be written as \code{~ z1 + z2 + ...}, where
\code{z1}, \code{z2}, ... are the excluded instruments for the two endogenous
variables in the asymmetric model: the average peers' outcomes and the average
difference in outcomes between an agent and her higher-performing peers. If
omitted, default instruments are generated using \link{gen.instrument}.}

\item{Glist}{The adjacency matrix or list of adjacency matrices. For networks
consisting of multiple subnets (e.g., schools), \code{Glist} must be a list,
where the \code{s}-th element is an \eqn{n_s x n_s} adjacency matrix and
\eqn{n_s} is the number of nodes in subnet \code{s}.}

\item{data}{An optional data frame, list, or environment (or an object that can
be coerced to a data frame via \link[base]{as.data.frame}) containing the
variables in the model. If a variable is not found in \code{data}, it is
retrieved from \code{environment(formula)}, typically the environment from
which \code{asypeer.estim} is called.}

\item{common.gamma}{A character vector indicating the control variables assumed
to have the same marginal effect on marginal utility for isolated and
non-isolated agents. If omitted, all control variables are included. A value
of \code{NULL} or \code{character(0)} indicates that none are included.}

\item{spillover}{A logical value indicating whether the model includes an
additional spillover component alongside the asymmetric conformity effect.}

\item{asymmetry}{A logical value indicating whether preferences for conformity
are asymmetric.}

\item{weight}{A character string specifying the weighting matrix used in GMM
estimation. Available options are \code{"identity"} for GMM with the identity
matrix, \code{"IV"} for the standard instrumental-variable GMM estimator, and
\code{"optimal"} for GMM with the optimal weighting matrix.}

\item{HAC}{A character string specifying the correlation structure of the
idiosyncratic errors used for covariance estimation. Options are
\code{"iid"} for independent errors; \code{"group iid"} for independence within
groups of isolated and non-isolated agents; \code{"hetero"} for
heteroskedastic but non-autocorrelated errors; \code{"cluster"} for
heteroskedastic errors with within-subnetwork correlation; and
\code{"cboot"} for the pair-bootstrap version of \code{"cluster"}.}

\item{nboot}{The number of bootstrap replications when \code{HAC = "cluster"}.}

\item{fixed.effects}{A logical value indicating whether subnetwork fixed effects
are included in the model.}

\item{tol}{A numeric tolerance used in QR factorization to detect collinear
columns in the matrices of explanatory variables and instruments, ensuring a
full-rank matrix (see \link[base]{qr}).}

\item{nthread}{The number of CPU cores (threads) used for parallel computation.}

\item{drop}{A logical vector of the same length as the sample, indicating which
observations should be dropped. This can be used, for example, to remove false
isolates or to estimate the model only for non-isolated agents. These
observations cannot be removed from the network structure because they may
remain connected to other agents.}

\item{tol.optim}{A numeric tolerance used in \code{\link[stats]{optimize}}. The objective
function is concentrated in a function of \eqn{\beta_l} or \eqn{\beta} alone,
which is optimized using \link[stats]{optimize}.}

\item{gen.inst.arg}{A list of arguments for the \code{\link{gen.instrument}}
function, used when instruments are to be generated. These arguments exclude
\code{formula}, \code{Glist}, \code{data}, \code{asymmetry},
\code{fold.nthread}, \code{drop}, and \code{tol}, which are taken directly
from the \code{asypeer.estim} function.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A list containing:
\item{model.info}{A list with information about the model, such as the number of subnets, number of observations, and other key details.}
\item{gmm}{A list of GMM estimation results, including parameter estimates, the covariance matrix, and related statistics.}
\item{data}{A list including the original data used to estimate the model, as well as the endogenous variables and their instruments.}
}
\description{
\code{asypeer.estim} estimates the asymmetric peer effects model. The instruments are generated automatically following the procedure described
in the paper and implemented in \link{gen.instruments}. The user can also supply her own instruments.
}
\examples{
if (requireNamespace("PartialNetwork", quietly = TRUE)) {
library(PartialNetwork)
ngr  <- 50  # Number of subnets
nvec <- rep(30, ngr)  # Size of subnets
n    <- sum(nvec)

### Simulating Data
## Network matrix
G   <- lapply(1:ngr, function(z) {
 Gz <- matrix(rbinom(nvec[z]^2, 1, 0.3), nvec[z], nvec[z])
 diag(Gz) <- 0
 # Adding isolated nodes (important for the structural model)
 niso <- sample(0:nvec[z], 1, prob = ((nvec[z] + 1):1)^5 / sum(((nvec[z] + 1):1)^5))
 if (niso > 0) {
   Gz[sample(1:nvec[z], niso), ] <- 0
 }
 Gz
})

Gnorm   <- norm.network(G)
X       <- cbind(rnorm(n, 0, 2), rpois(n, 2))
GX      <- peer.avg(Gnorm, X)
delta   <- 0.25
beta    <- c(0.3, 0.6)
gamma   <- c(4, 1, -0.7, 0, -0.5) 
eps     <- rnorm(n, 0, 0.5) 

## Generating `y`
y <- asypeer.sim(formula = ~ X + GX, Glist = Gnorm, delta = delta, beta = beta, 
                gamma = gamma, epsilon = eps)
y <- y$y

### Estimating the asymmetric peer effects model
est <- asypeer.estim(formula=y ~ X + GX, Glist = Gnorm, spillover = TRUE)
summary(est, diagnostic = TRUE)
}                  
}
