% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instruments.R
\name{gen.instrument}
\alias{gen.instrument}
\alias{gen.instruments}
\alias{gen.insts}
\alias{gen.inst}
\title{Generate Instruments for the Asymmetric Peer Effects Model}
\usage{
gen.instrument(
  formula,
  Glist,
  data,
  asymmetry = TRUE,
  estimator = c("ols", "logit"),
  power = c(1, 1),
  full = FALSE,
  nfold = 2,
  checkrank = TRUE,
  tol = 1e-10,
  nthread = 1,
  drop = NULL,
  ...
)

gen.instruments(
  formula,
  Glist,
  data,
  asymmetry = TRUE,
  estimator = c("ols", "logit"),
  power = c(1, 1),
  full = FALSE,
  nfold = 2,
  checkrank = TRUE,
  tol = 1e-10,
  nthread = 1,
  drop = NULL,
  ...
)

gen.insts(
  formula,
  Glist,
  data,
  asymmetry = TRUE,
  estimator = c("ols", "logit"),
  power = c(1, 1),
  full = FALSE,
  nfold = 2,
  checkrank = TRUE,
  tol = 1e-10,
  nthread = 1,
  drop = NULL,
  ...
)

gen.inst(
  formula,
  Glist,
  data,
  asymmetry = TRUE,
  estimator = c("ols", "logit"),
  power = c(1, 1),
  full = FALSE,
  nfold = 2,
  checkrank = TRUE,
  tol = 1e-10,
  nthread = 1,
  drop = NULL,
  ...
)
}
\arguments{
\item{formula}{An object of class \link[stats]{formula}: a symbolic description
of the model. The formula should be specified as \code{y ~ x1 + x2}, where
\code{y} is the outcome variable and \code{x1}, \code{x2}, ... are control
variables, which may include contextual variables such as peer averages.}

\item{Glist}{The adjacency matrix or a list of adjacency matrices. For networks
composed of multiple subnets (e.g., schools), \code{Glist} must be a list,
where the \code{s}-th element is an \eqn{n_s \times n_s} adjacency matrix,
and \eqn{n_s} is the number of nodes in subnet \code{s}.}

\item{data}{An optional data frame, list, or environment (or an object that can
be coerced to a data frame via \link[base]{as.data.frame}) containing the
variables in the model. If a variable is not found in \code{data}, it is
searched for in \code{environment(formula)}, typically the environment from
which \code{gen.instrument} is called.}

\item{asymmetry}{A logical value indicating whether preferences for conformity
are asymmetric.}

\item{estimator}{A character vector of length 2 specifying the estimators used
for the intensive margin \eqn{E(y_j - y_i \mid y_j - y_i > 0)} (or \eqn{E(\bar{y}_i}))
and the extensive margin \eqn{P(y_j > y_i)}.

Supported parametric options include \code{"ols"} and \code{"lasso"} for linear
models, and \code{"glm"} for generalized linear models (only for the extensive
margin). Nonparametric options include \code{"rf"} (random forest) and
\code{"xgboost"} (gradient boosting).

For example, \code{estimator = c("xgboost", "glm")} uses gradient boosting for
the intensive margin and a generalized linear model for the extensive margin.}

\item{power}{A numeric vector of length 2 indicating the maximum walk lengths
(typically \eqn{k} in \eqn{G^k X}) used to construct or estimate instruments for
\eqn{\bar{y}_i} and \eqn{\check{y}_i}. The two entries allow different values
of \eqn{k} for each endogenous variable.}

\item{full}{A logical value. If \code{TRUE}, predictions for both
\eqn{\bar{y}_i} and \eqn{\check{y}_i} are used as instruments. If \code{FALSE},
only the prediction for \eqn{\check{y}_i} is used as an instrument, while
\eqn{\bar{y}_i} is instrumented using the usual instruments, \eqn{G^k X}.}

\item{nfold}{A strictly positive integer specifying the number of folds used for
cross-fitting in the estimation of the probability model.}

\item{checkrank}{A logical value indicating whether linearly dependent columns
in the matrix of generated instruments should be dropped.}

\item{tol}{A numeric tolerance used in QR factorization to detect linearly
dependent columns in the matrices of explanatory variables and instruments,
ensuring a full-rank matrix (see \link[base]{qr}).}

\item{nthread}{Number of CPU cores (threads) used to run parts of the estimation
in parallel.}

\item{drop}{A logical vector of the same length as the sample, indicating which
observations should be dropped. This can be used, for example, to remove false
isolates or to estimate the model only on non-isolated agents. These
observations cannot be physically removed from the network structure because
they may still be connected to other agents.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A list containing:
\item{model.info}{A list with information about the model, such as the estimator, the number of folds, and other key details.}
\item{instruments}{A matrix of generated instruments.}
}
\description{
\code{gen.instrument} generates instruments for the endogenous variables in asymmetric peer effects models.
}
\examples{
if (requireNamespace("PartialNetwork", quietly = TRUE)) {
library(PartialNetwork)
ngr  <- 50  # Number of subnets
nvec <- rep(30, ngr)  # Size of subnets
n    <- sum(nvec)

### Simulating Data
## Network matrix
G <- lapply(1:ngr, function(z) {
  Gz       <- matrix(rbinom(nvec[z]^2, 1, 0.3), nvec[z], nvec[z])
  diag(Gz) <- 0
  # Adding isolated nodes (important for the structural model)
  niso <- sample(0:nvec[z], 1, prob = ((nvec[z] + 1):1)^5 / sum(((nvec[z] + 1):1)^5))
  if (niso > 0) {
    Gz[sample(1:nvec[z], niso), ] <- 0
  }
  Gz
})
Gnorm   <- norm.network(G)
X       <- cbind(rnorm(n, 0, 2), rpois(n, 2))
GX      <- peer.avg(Gnorm, X)
delta   <- 0.25
beta    <- c(0.3, 0.6)
gamma   <- c(4, 1, -0.7, 0, -0.5) 
eps     <- rnorm(n, 0, 0.5) 

## Generating `y`
y <- asypeer.sim(formula = ~ X + GX, Glist = Gnorm, delta = delta, 
                 beta = beta, gamma = gamma, epsilon = eps)
y <- y$y

### Generating instruments
ins <- gen.instrument(formula = y ~ X, Glist = Gnorm, 
                      estimator = c("ols", "logit"))}
 
}
