% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spillover.R
\name{plot.spillover}
\alias{plot.spillover}
\title{Plotting the effect of a targeted intervention}
\usage{
\method{plot}{spillover}(x, metric = "all", ...)
}
\arguments{
\item{x}{An object of class \code{\link{spillover}} as returned by the function \link{spillover}.}

\item{metric}{A character string specifying which plots should be generated.
Available options include \code{"spillover"}, \code{"gain"}, \code{"outcome"}
and any combination of these options.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
\code{plot.spillover} creates plots that illustrate the spillovers across different
intervention's budgets and targeting methods.
}
