% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spillover.R
\name{plot.spillover}
\alias{plot.spillover}
\title{Plot the Effects of a Targeted Intervention}
\usage{
\method{plot}{spillover}(x, metric = c("spillover", "outcome"), range, ...)
}
\arguments{
\item{x}{An object of class \code{\link{spillover}}, as returned by
\link{spillover}.}

\item{metric}{A character string specifying which metrics should be plotted.
Available options are \code{"spillover"}, \code{"gain"},
\code{"outcome"}, or any combination of these.}

\item{range}{A vector of integer indices indicating the ranks to be
considered.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
\code{plot.spillover} generates plots illustrating spillovers under different
intervention budgets and targeting methods.
}
