% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asypeer.estim.R
\name{summary.asypeer.estim}
\alias{summary.asypeer.estim}
\alias{print.summary.asypeer.estim}
\alias{print.asypeer.estim}
\title{Summary and Print Methods for the Asymmetric Peer Effects Model}
\usage{
\method{summary}{asypeer.estim}(
  object,
  diagnostic = FALSE,
  diagnostics = FALSE,
  SW = FALSE,
  KP = diagnostics || diagnostic,
  nthread = 1L,
  ...
)

\method{print}{summary.asypeer.estim}(x, ...)

\method{print}{asypeer.estim}(x, ...)
}
\arguments{
\item{object}{An object of class \code{\link{asypeer.estim}} as returned by the function \link{asypeer.estim}.}

\item{diagnostics, diagnostic}{A logical value indicating whether diagnostic tests for the IV GMM should be performed.
These include an F-test of the first-stage regression for weak instruments, a Wu-Hausman test
for endogeneity, and a Hansen's J-test for overidentifying restrictions (the latter only when
the number of instruments exceeds the number of regressors).}

\item{SW}{Logical value indicating whether the Sanderson–Windmeijer conditional F-statistic is computed instead of the classical first-stage F-statistic.}

\item{KP}{A logical value indicating whether a Kleibergen-Paap Wald test should be performed in addition to the standard F test
of the first-stage regression for weak instruments. should be performed instead of the standard F test.}

\item{nthread}{A strictly positive integer specifying the number of threads used in
computationally intensive steps of the estimation procedure.}

\item{...}{Further arguments passed to or from other methods.}

\item{x}{An object of class \code{\link{summary.asypeer.estim}} or \code{\link{asypeer.estim}} as returned by the function \link{summary.asypeer.estim} or \link{asypeer.estim}, respectively.}
}
\value{
A list containing:
\item{model.info}{A list containing information about the model, such as the number of subnets,
number of observations, and other key details.}
\item{coefficients}{A summary of coefficient estimates, standard errors, and p-values.}
\item{diagnostics}{A summary of the diagnostic tests for the instrumental-variable regression,
if requested.}
\item{gmm}{A list of GMM estimation results, including parameter estimates, the covariance matrix,
and related statistics.}
}
\description{
Summary and print methods for objects of class \code{\link{asypeer.estim}}.
}
